refactor: change grade's type to Double instead of Integer

+++ src/main/java/org/demo/academicsystem/entity/Grade.java
@@ -23,7 +23,7 @@ public class Grade {
     private Long id;
 
     @Column(nullable = false)
-    private Integer grade;
+    private Double grade;
 
     @Column(length = 100)
     private String comments;
